feat: refactor core utility form
