+++ src/View/Helper/SchemaHelper.php
@@ -13,7 +13,9 @@
 
 namespace App\View\Helper;
 
-use App\Core\Utility\Form;
+use App\Core\Utility\Form\Control;
+use App\Core\Utility\Form\ControlType;
+use App\Core\Utility\Form\Options;
 use Cake\Utility\Hash;
 use Cake\Utility\Inflector;
 use Cake\View\Helper;
@@ -42,13 +44,13 @@ class SchemaHelper extends Helper
      */
     public function controlOptions(string $name, $value, $schema = []): array
     {
-        $options = Form::customControlOptions($name, $value);
+        $options = Options::customControl($name, $value);
         if (!empty($options)) {
             return $options;
         }
-        $type = Form::controlTypeFromSchema((array)$schema);
+        $type = ControlType::fromSchema((array)$schema);
 
-        return Form::control((array)$schema, $type, $value);
+        return Control::control((array)$schema, $type, $value);
     }
 
     /**
@@ -135,7 +137,7 @@ class SchemaHelper extends Helper
                 return static::typeFromSchema($subSchema);
             }
         }
-        if (empty($schema['type']) || !in_array($schema['type'], Form::SCHEMA_PROPERTY_TYPES)) {
+        if (empty($schema['type']) || !in_array($schema['type'], ControlType::SCHEMA_PROPERTY_TYPES)) {
             return 'string';
         }
         $format = Hash::get($schema, 'format');
