--- tests/TestCase/Core/Utility/FormTest.php
@@ -1,618 +1,0 @@
-<?php
-/**
- * BEdita, API-first content management framework
- * Copyright 2018 ChannelWeb Srl, Chialab Srl
- *
- * This file is part of BEdita: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published
- * by the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * See LICENSE.LGPL or <http://gnu.org/licenses/lgpl-3.0.html> for more details.
- */
-
-namespace App\Test\TestCase\Core\Utility;
-
-use App\Core\Utility\Form;
-use Cake\Core\Configure;
-use Cake\TestSuite\TestCase;
-
-/**
- * {@see \App\Core\Utility\Form} Test Case
- *
- * @coversDefaultClass \App\Core\Utility\Form
- */
-class FormTest extends TestCase
-{
-    /**
-     * Data provider for `testControlTypeFromSchema` test case.
-     *
-     * @return array
-     */
-    public function controlTypeFromSchemaProvider(): array
-    {
-        return [
-            'string' => [
-                'text',
-                [
-                    'type' => 'string',
-                ],
-            ],
-            'html' => [
-                'richtext',
-                [
-                    'type' => 'string',
-                    'contentMediaType' => 'text/html',
-                ],
-            ],
-            'txt' => [
-                'plaintext',
-                [
-                    'type' => 'string',
-                    'contentMediaType' => 'text/plain',
-                ],
-            ],
-            'date-time' => [
-                'date-time',
-                [
-                    'type' => 'string',
-                    'format' => 'date-time',
-                ],
-            ],
-            'date' => [
-                'date',
-                [
-                    'type' => 'string',
-                    'format' => 'date',
-                ],
-            ],
-            'integer' => [
-                'number',
-                [
-                    'type' => 'integer',
-                ],
-            ],
-            'number' => [
-                'number',
-                [
-                    'type' => 'number',
-                ],
-            ],
-            'checkbox' => [
-                'checkbox',
-                [
-                    'type' => 'boolean',
-                ],
-            ],
-            'one of' => [
-                'checkbox',
-                [
-                    'type' => 'string',
-                    'oneOf' => [
-                        [
-                            'type' => 'null',
-                        ],
-                        [
-                            'type' => 'boolean',
-                        ],
-                    ],
-                ],
-            ],
-            'not an array' => [
-                'text',
-                false,
-            ],
-            'no type' => [
-                'text',
-                [
-                    'const' => 13,
-                ],
-            ],
-            'json' => [
-                'json',
-                [
-                    'type' => 'object',
-                ],
-            ],
-            'unknown' => [
-                'text',
-                [
-                    'type' => 'unknown',
-                ],
-            ],
-            'enum' => [
-                'enum',
-                [
-                    'type' => 'string',
-                    'enum' => ['a', 'b', 'c'],
-                ],
-            ],
-            'array' => [
-                'checkbox',
-                [
-                    'type' => 'array',
-                    'oneOf' => [
-                        [
-                            'type' => 'null',
-                        ],
-                        [
-                            'type' => 'array',
-                            'uniqueItems' => true,
-                            'items' => [
-                                'type' => 'string',
-                                'enum' => ['a', 'b', 'c', 'd'],
-                            ],
-                        ],
-                    ],
-                ],
-            ],
-            'categories' => [
-                'categories',
-                [
-                    'type' => 'categories',
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * Test `controlTypeFromSchema()` method.
-     *
-     * @param string $expected Expected result.
-     * @param array|bool $schema Schema.
-     * @return void
-     *
-     * @dataProvider controlTypeFromSchemaProvider()
-     * @covers ::controlTypeFromSchema()
-     * @covers ::typeFromString()
-     * @covers ::typeFromNumber()
-     * @covers ::typeFromInteger()
-     * @covers ::typeFromBoolean()
-     * @covers ::typeFromArray()
-     * @covers ::typeFromObject()
-     */
-    public function testControlTypeFromSchema(string $expected, $schema): void
-    {
-        $actual = Form::controlTypeFromSchema($schema);
-
-        static::assertSame($expected, $actual);
-    }
-
-    /**
-     * Data provider for `testCustomControlOptions` test case.
-     * Custom control types (@see Form::customControls):
-     *
-     *     'lang', 'status', 'old_password', 'password', 'confirm-password', 'title'
-     *
-     * @return array
-     */
-    public function customControlOptionsProvider(): array
-    {
-        return [
-            'not custom' => [
-                'name',
-                null,
-                [],
-            ],
-            'lang' => [
-                'lang',
-                'en',
-                [
-                    'type' => 'text',
-                    'value' => 'en',
-                ],
-            ],
-            'lang options' => [
-                'lang',
-                'en',
-                [
-                    'type' => 'select',
-                    'options' => [
-                        [
-                            'value' => 'en',
-                            'text' => 'English',
-                        ],
-                        [
-                            'value' => 'de',
-                            'text' => 'Deutsch',
-                        ],
-                    ],
-                    'value' => 'en',
-                ],
-                [
-                    'Project.config.I18n.languages' => [
-                        'en' => 'English',
-                        'de' => 'Deutsch',
-                    ],
-                ],
-            ],
-            'status' => [
-                'status',
-                'draft',
-                [
-                    'type' => 'radio',
-                    'options' => [
-                        ['value' => 'on', 'text' => __('On')],
-                        ['value' => 'draft', 'text' => __('Draft')],
-                        ['value' => 'off', 'text' => __('Off')],
-                    ],
-                    'templateVars' => [
-                        'containerClass' => 'status',
-                    ],
-                    'value' => 'draft',
-                ],
-            ],
-            'password' => [
-                'password',
-                '12345',
-                [
-                    'class' => 'password',
-                    'placeholder' => __('new password'),
-                    'autocomplete' => 'new-password',
-                    'default' => '',
-                    'value' => '12345',
-                ],
-            ],
-            'confirm-password' => [
-                'confirm-password',
-                '12345',
-                [
-                    'label' => __('Retype password'),
-                    'id' => 'confirm_password',
-                    'name' => 'confirm-password',
-                    'class' => 'confirm-password',
-                    'placeholder' => __('confirm password'),
-                    'autocomplete' => 'new-password',
-                    'default' => '',
-                    'type' => 'password',
-                    'value' => '12345',
-                ],
-            ],
-            'title' => [
-                'title',
-                'dummy',
-                [
-                    'class' => 'title',
-                    'type' => 'text',
-                    'templates' => [
-                        'inputContainer' => '<div class="input title {{type}}{{required}}">{{content}}</div>',
-                    ],
-                    'value' => 'dummy',
-                ],
-            ],
-            'start_date' => [
-                'start_date',
-                '2020-10-14',
-                [
-                    'type' => 'text',
-                    'v-datepicker' => 'true',
-                    'date' => 'true',
-                    'time' => 'true',
-                    'value' => '2020-10-14',
-                    'templates' => [
-                        'inputContainer' => '<div class="input datepicker {{type}}{{required}}">{{content}}</div>',
-                    ],
-                ],
-            ],
-            'end_date' => [
-                'end_date',
-                '2020-10-15',
-                [
-                    'type' => 'text',
-                    'v-datepicker' => 'true',
-                    'date' => 'true',
-                    'time' => 'true',
-                    'value' => '2020-10-15',
-                    'templates' => [
-                        'inputContainer' => '<div class="input datepicker {{type}}{{required}}">{{content}}</div>',
-                    ],
-                ],
-            ],
-            'date_ranges' => [
-                'date_ranges',
-                null,
-                [
-                    'type' => 'text',
-                    'v-datepicker' => 'true',
-                    'date' => 'true',
-                    'time' => 'true',
-                    'value' => null,
-                    'templates' => [
-                        'inputContainer' => '<div class="input datepicker {{type}}{{required}}">{{content}}</div>',
-                    ],
-                ],
-            ],
-            'coords' => [
-                'coords',
-                'POINT(11.123125 44.6123245)',
-                [
-                    'class' => 'coordinates',
-                    'templates' => [
-                        'inputContainer' => '<div class="input coordinates {{type}}{{required}}">{{content}}<coordinates-view latitude="44.6123245" longitude="11.123125" /></div>',
-                    ],
-                    'type' => 'readonly',
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * Test `customControlOptions` method.
-     *
-     * @param string $name The field name.
-     * @param mixed|null $value The field value.
-     * @param array $expected Expected result.
-     * @param array $config Configuration.
-     * @return void
-     *
-     * @dataProvider customControlOptionsProvider()
-     * @covers ::customControlOptions()
-     * @covers ::langOptions
-     * @covers ::dateRangesOptions(()
-     * @covers ::startDateOptions()
-     * @covers ::endDateOptions()
-     * @covers ::statusOptions
-     * @covers ::oldpasswordOptions
-     * @covers ::passwordOptions
-     * @covers ::confirmpasswordOptions
-     * @covers ::titleOptions
-     * @covers ::coordsOptions
-     * @covers ::typeFromString
-     * @covers ::typeFromNumber
-     * @covers ::typeFromInteger
-     * @covers ::typeFromBoolean
-     * @covers ::typeFromArray
-     * @covers ::typeFromObject
-     */
-    public function testCustomControlOptions(string $name, $value, array $expected, array $config = []): void
-    {
-        if (!empty($config)) {
-            Configure::write($config);
-        }
-        $actual = Form::customControlOptions($name, $value);
-        ksort($expected);
-        static::assertSame($expected, $actual);
-    }
-
-    /**
-     * Data provider for `testControl` test case.
-     *
-     * @return array
-     */
-    public function controlProvider(): array
-    {
-        $value = 'something';
-
-        return [
-            'other' => [
-                [], // schema
-                'other', // type
-                $value, // value
-                ['type' => 'other'] + compact('value'), // expected
-            ],
-            'json' => [
-                [],
-                'json',
-                $value,
-                [
-                    'type' => 'textarea',
-                    'v-jsoneditor' => 'true',
-                    'class' => 'json',
-                    'value' => json_encode($value),
-                ],
-            ],
-            'richtext' => [
-                [],
-                'richtext',
-                $value,
-                [
-                    'type' => 'textarea',
-                    'v-richeditor' => '""',
-                    'value' => $value,
-                ],
-            ],
-            'plaintext' => [
-                [],
-                'plaintext',
-                $value,
-                [
-                    'type' => 'textarea',
-                    'value' => $value,
-                ],
-            ],
-            'date-time' => [
-                [],
-                'date-time',
-                '2018-01-31 08:45:00',
-                [
-                    'type' => 'text',
-                    'v-datepicker' => 'true',
-                    'date' => 'true',
-                    'time' => 'true',
-                    'value' => '2018-01-31 08:45:00',
-                    'templates' => [
-                        'inputContainer' => '<div class="input datepicker {{type}}{{required}}">{{content}}</div>',
-                    ],
-                ],
-            ],
-            'date' => [
-                [],
-                'date',
-                '2018-01-31',
-                [
-                    'type' => 'text',
-                    'v-datepicker' => 'true',
-                    'date' => 'true',
-                    'value' => '2018-01-31',
-                    'templates' => [
-                        'inputContainer' => '<div class="input datepicker {{type}}{{required}}">{{content}}</div>',
-                    ],
-                ],
-            ],
-            'checkbox' => [
-                [
-                    'type' => 'array',
-                    'oneOf' => [
-                        [
-                            'type' => 'null',
-                        ],
-                        [
-                            'type' => 'array',
-                            'uniqueItems' => true,
-                            'items' => [
-                                'type' => 'string',
-                                'enum' => ['a', 'b', 'c', 'd'],
-                            ],
-                        ],
-                    ],
-                ],
-                'checkbox',
-                null,
-                [
-                    'type' => 'select',
-                    'options' => [
-                        [
-                            'value' => 'a',
-                            'text' => 'A',
-                        ],
-                        [
-                            'value' => 'b',
-                            'text' => 'B',
-                        ],
-                        [
-                            'value' => 'c',
-                            'text' => 'C',
-                        ],
-                        [
-                            'value' => 'd',
-                            'text' => 'D',
-                        ],
-                    ],
-                    'multiple' => 'checkbox',
-                ],
-            ],
-            'enum' => [
-                [
-                    'type' => 'string',
-                    'enum' => [
-                        'good',
-                        'bad',
-                    ],
-                ],
-                'enum',
-                'good',
-                [
-                    'type' => 'select',
-                    'options' => [
-                        ['value' => 'good', 'text' => 'Good'],
-                        ['value' => 'bad', 'text' => 'Bad'],
-                    ],
-                ],
-            ],
-            'enum nullable' => [
-                [
-                    'oneOf' => [
-                        [
-                            'type' => 'null',
-                        ],
-                        [
-                            'type' => 'string',
-                            'enum' => [
-                                'good',
-                                'bad',
-                            ],
-                        ],
-                    ],
-                ],
-                'enum',
-                'good',
-                [
-                    'type' => 'select',
-                    'options' => [
-                        ['value' => '', 'text' => ''],
-                        ['value' => 'good', 'text' => 'Good'],
-                        ['value' => 'bad', 'text' => 'Bad'],
-                    ],
-                ],
-            ],
-        ];
-    }
-
-    /**
-     * Test `control` method
-     *
-     * @param array $schema Object schema array.
-     * @param string $type Control type.
-     * @param string|null $value Property value.
-     * @param array $expected The expected control.
-     * @return void
-     *
-     * @dataProvider controlProvider()
-     * @covers ::control
-     * @covers ::jsonControl
-     * @covers ::richtextControl
-     * @covers ::plaintextControl
-     * @covers ::datetimeControl
-     * @covers ::dateControl
-     * @covers ::checkboxControl
-     * @covers ::enumControl
-     */
-    public function testControl(array $schema, string $type, ?string $value, array $expected): void
-    {
-        $actual = Form::control($schema, $type, $value);
-
-        static::assertSame($expected, $actual);
-    }
-
-    /**
-     * Data provider for `testGetMethod` test case.
-     *
-     * @return array
-     */
-    public function getMethodProvider(): array
-    {
-        return [
-            'name with chars to remove' => [
-                'old_passwordOptions',
-                [Form::class, 'oldPasswordOptions'],
-            ],
-            'name with chars to remove' => [
-                'confirm_passwordOptions',
-                [Form::class, 'confirmPasswordOptions'],
-            ],
-        ];
-    }
-
-    /**
-     * Test `getMethod` method
-     *
-     * @param string $methodName The method name
-     * @param array $expected The expected method array
-     * @return void
-     *
-     * @dataProvider getMethodProvider()
-     * @covers ::getMethod
-     */
-    public function testGetMethod(string $methodName, array $expected): void
-    {
-        $actual = Form::getMethod($methodName);
-
-        static::assertSame($expected, $actual);
-    }
-
-    /**
-     * Test `getMethod` method exception 'not callable'
-     *
-     * @return void
-     *
-     * @covers ::getMethod
-     */
-    public function testGetMethodNotCallable(): void
-    {
-        $methodName = 'dummy';
-        $expected = new \InvalidArgumentException(sprintf('Method "%s" is not callable', $methodName));
-        static::expectException(get_class($expected));
-        static::expectExceptionMessage($expected->getMessage());
-        Form::getMethod($methodName);
-    }
-}
